Avaliação 2

+++ src/Pontos.java
@@ -1,26 +1,26 @@
 public class Pontos {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Pontos(int x, int y) {
+    public Pontos(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return this.x;
     }
 
-    public int getY() {
+    public double getY() {
         return this.y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
